Renamed some files + added import statements to __init__.py files so we can drop .datasource in config.py
