+++ app/connectors_service/connectors/config.py
@@ -114,19 +114,19 @@ def _default_config():
             "azure_blob_storage": "connectors.sources.azure_blob_storage:AzureBlobStorageDataSource",
             "box": "connectors.sources.box:BoxDataSource",
             "confluence": "connectors.sources.confluence:ConfluenceDataSource",
-            "dir": "connectors.sources.directory.datasource:DirectoryDataSource",
-            "dropbox": "connectors.sources.dropbox.datasource:DropboxDataSource",
+            "dir": "connectors.sources.directory:DirectoryDataSource",
+            "dropbox": "connectors.sources.dropbox:DropboxDataSource",
             "github": "connectors.sources.github:GitHubDataSource",
-            "gmail": "connectors.sources.gmail.datasource:GMailDataSource",
-            "google_cloud_storage": "connectors.sources.google_cloud_storage.datasource:GoogleCloudStorageDataSource",
-            "google_drive": "connectors.sources.google_drive.datasource:GoogleDriveDataSource",
+            "gmail": "connectors.sources.gmail:GMailDataSource",
+            "google_cloud_storage": "connectors.sources.google_cloud_storage:GoogleCloudStorageDataSource",
+            "google_drive": "connectors.sources.google_drive:GoogleDriveDataSource",
             "graphql": "connectors.sources.graphql:GraphQLDataSource",
             "jira": "connectors.sources.jira:JiraDataSource",
             "microsoft_teams": "connectors.sources.microsoft_teams:MicrosoftTeamsDataSource",
-            "mongodb": "connectors.sources.mongo.datasource:MongoDataSource",
+            "mongodb": "connectors.sources.mongo:MongoDataSource",
             "mssql": "connectors.sources.mssql:MSSQLDataSource",
             "mysql": "connectors.sources.mysql:MySqlDataSource",
-            "network_drive": "connectors.sources.network_drive.datasource:NASDataSource",
+            "network_drive": "connectors.sources.network_drive:NASDataSource",
             "notion": "connectors.sources.notion:NotionDataSource",
             "onedrive": "connectors.sources.onedrive:OneDriveDataSource",
             "oracle": "connectors.sources.oracle:OracleDataSource",
