+++ app/connectors_service/connectors/sources/directory/__init__.py
@@ -3,3 +3,4 @@
 # or more contributor license agreements. Licensed under the Elastic License 2.0;
 # you may not use this file except in compliance with the Elastic License 2.0.
 #
+from connectors.sources.directory.datasource import DirectoryDataSource
